# Add reference fallback for TARGET_ARCH!=hifimini