A follow-up change will add in calls to the XaNNLib to get an optimized softmax implementation.

Tested that keyword_benchmark has a latency increase of ~1000 ticks for the Fusion F1.

This latency improvement does not matter since the current change is setting the stage for pulling in the optimized implementation.

This command:
```
make -f tensorflow/lite/micro/tools/make/Makefile TARGET=xtensa OPTIMIZED_KERNEL_DIR=xtensa TARGET_ARCH=fusion_f1 XTENSA_CORE=F1_190305_swupgrade test_keyword_benchmark -j8
```

before this change:
```
InitializeKeywordRunner() took 279548 ticks (279 ms)
KeywordRunNIerations(1) took 151249 ticks (151 ms)
KeywordRunNIerations(10) took 1511997 ticks (1511 ms)
```

after this change:
```
InitializeKeywordRunner() took 201464 ticks (201 ms)
KeywordRunNIerations(1) took 152158 ticks (152 ms)
KeywordRunNIerations(10) took 1521087 ticks (1521 ms)
```

Progress towards http://b/177457688